Now use the httlib2 instead of urllib2. This library is under the MIT license, homepage is : http://code.google.com/p/httplib2/
Major feature of httplib2 is the use of persistent connexions
for HTTP requests so it reduce the load on the webserver and
we can send several requests faster.
Httplib2 has some other interesting features, for exemple use
of proxies and the possibility to use a SOCKS proxy.

The main disadvantage is the lack of cookies management so we will
have to play with the cookielib and see how we can inject the Cookie
headers in the HTTP requests

We should also take a look at error handling (for example, see if the
"HTTP 500 error code" bug is fixed with httplib2 and how redirections
are managed)

Wapiti is fasteeeeeeeeeeer !! :)